+++ app/Enums/Server/VirtualizationType.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Convoy\Enums\Server;
+
+enum VirtualizationType: string
+{
+    case KVM = 'kvm';
+    case LXC = 'lxc';
+}