+++ app/Http/Requests/Client/Servers/Settings/UpdateBiosTypeRequest.php
@@ -2,7 +2,7 @@
 
 namespace Convoy\Http\Requests\Client\Servers\Settings;
 
-use Convoy\Enums\Server\Cloudinit\BiosType;
+use Convoy\Enums\Server\BiosType;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rules\Enum;