+++ app/Jobs/Node/MonitorIsoDownloadJob.php
@@ -3,7 +3,7 @@
 namespace Convoy\Jobs\Node;
 
 use Convoy\Models\ISO;
-use Convoy\Services\Nodes\IsoLibrary\IsoMonitorService;
+use Convoy\Services\Nodes\Isos\IsoMonitorService;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldBeUnique;
 use Illuminate\Contracts\Queue\ShouldQueue;