+++ public/index.php
@@ -2,6 +2,7 @@
 
 
 
+
 use Illuminate\Contracts\Http\Kernel;
 use Illuminate\Http\Request;